perf: uint8 pool + conv channel-chunking to unblock 8000 gates / conv-MNIST OOM
精度でなくインフラのレバー。VRAM(RTX 3060 6GB)で詰んでいた2実験を解消:
- 常駐プールを float32 -> uint8(ハードビットは厳密に0/1で可逆・ビット等価。
  学習/hard評価の直前にバッチ/チャンク単位で float へ戻す)
- conv soft学習をチャンネルチャンク+チャンクごと勾配checkpointでピーク定数化

ビット等価: tests.py 全10ケース + 新設convピン(C64/tree3残差=0.6422@3)が
CUDAで厳密PASS、digits/MNISTプローブも実装前後で完全一致。convのマルチチャンクのみ
logits勾配のB×HW縮約順序が形状で変わり最終ビットの丸めが動きうる(相対~1e-6、
数学的に同値、既存ピン構成は全て1チャンクで完全等価/マルチチャンクは従来OOMで
走れなかった構成のみ発動)。

アンロック確認(max_memory_allocated):
- conv MNIST C32/tree3: 層1即死 -> 層1・2完走(ピーク3.59GiB)
- 8000+skip+residual+batch2048: 層3クラッシュ -> 完走(ピーク5.03GiB)。
  副産物: 20epoch短縮設定で94.19%@深さ3(champ94.27%@38に深さ3で肉薄)

+++ src/core.py
@@ -154,8 +154,10 @@ def next_pool(h, X, pool_prev, cfg):
 @torch.no_grad()
 def hard_batched(layer, x, budget=8192 * 500):  # bound the [B, G, 16] temporary
     chunk = max(1024, budget // layer.ia.numel())  # 500 gates -> 8192 rows (as before)
-    return layer.hard(x) if len(x) <= chunk else torch.cat(
-        [layer.hard(c) for c in x.split(chunk)])
+    # x はuint8プール(常駐メモリ削減、issue #9)でも可: チャンク単位でfloatへ
+    # 戻して評価する。float32入力では .float() は無コピーの恒等なので従来と同一
+    return layer.hard(x.float()) if len(x) <= chunk else torch.cat(
+        [layer.hard(c.float()) for c in x.split(chunk)])
 
 def reps(pool_w, cfg):
     """--recur: この層を何回反復適用するか。重み共有の再帰は配線次元が
